requirements.txt
Babel==2.4.0
Django==1.11.3
django-phonenumber-field==1.3.0
olefile==0.44
phonenumberslite==8.7.1
Pillow==4.2.1
pytz==2017.2
